Move slider to component dir
